Added a Queue System
Needed for MP3 player. Went for an object orientated approach. As this is new there isn't really any function, all it can do is exist, set a playlist to the queue which was used in main.py, and it displays the queue, index and playing track. I plan to use a circular queue or a similar data structure.